Almost done search functionality, fixing like functionnality that does not update in the homescreen + TODO remake the logic for gathering and saving liked properties in an array

+++ routes/properties.js
@@ -84,7 +84,7 @@ router.post("/like", async (req, res, next) => {
       if (isLiked) {
         query = `INSERT INTO Likes (id, propertyId) VALUES (${userId},${propertyId});`;
       } else {
-        query = `DELETE FROM Likes WHERE propertyId = ${propertyId};`;
+        query = `DELETE FROM Likes WHERE propertyId = ${propertyId} AND id = ${userId};`;
       }
       sql.connect(config, function (err) {
         if (err) {
@@ -111,6 +111,169 @@ router.post("/like", async (req, res, next) => {
     return res.status(403).json({ error: "Access denied." });
   }
 });
+
+router.post("/search", async (req, res, next) => {
+  const {
+    address,
+    bathRoomCounts,
+    bedRoomCounts,
+    cities,
+    kitchenRoomCounts,
+    maxPrice,
+    maxSize,
+    minPrice,
+    minSize,
+    parkingCounts,
+    propertyCategories,
+    propertyTypes,
+    searchText,
+  } = req.body;
+  const token = req.headers.authorization?.split(" ")[1];
+
+  // Decrypt the token and check the role in the payload
+  // Replace the decryption logic with your own implementation
+  const decryptedToken = decryptToken(token);
+  const userId = decryptedToken?.userId;
+  const role = decryptedToken?.role.toLowerCase();
+
+  role.trim();
+
+  if (role === "user" || role === "admin") {
+    // If the role is "user", proceed with fetching the dashboard data
+    const config = {
+      user: "homico_admin",
+      password: "Nejamaistrahir1997",
+      server: "homicoserver.database.windows.net",
+      database: "homico",
+      options: {
+        encrypt: true,
+      },
+    };
+    const pool = await sql.connect(config);
+
+    const address = ""; // Provide the desired address pattern
+    const bathroomCount = bathRoomCounts[0];
+    const bedroomCount = bedRoomCounts[0];
+    const cities = ""; // Provide the desired city name pattern
+    const kitchenRoomCount = kitchenRoomCounts[0];
+    const maxPricee = maxPrice;
+    const maxSizee = maxSize;
+    const minPricee = minPrice;
+    const minSizee = minSize;
+    const parkingCountse = parkingCounts[0];
+    const propertyTypese = ""; // Provide the desired property type pattern
+    const searchTexte = searchText; // Provide the desired search text pattern
+
+    let query;
+
+    if (userId) {
+      query = `
+    SELECT
+        property.*,
+        city.*,
+        PropertyType.*,
+        CASE WHEN Likes.propertyId IS NOT NULL THEN 1 ELSE 0 END AS is_liked
+    FROM
+        property
+    JOIN
+        city ON property.cityId = city.cityId
+    JOIN
+        PropertyType ON property.propertyTypeId = PropertyType.propertyTypeId
+    LEFT JOIN
+        Likes ON property.propertyId = Likes.propertyId AND Likes.id = @user
+    WHERE
+        property.addres LIKE '%' + @address + '%'
+        AND property.baths > @bathroomCount
+        AND property.beds > @bedroomCount
+        AND city.nom LIKE '%' + @cities + '%'
+        AND property.kitchens > @kitchenRoomCount
+        AND property.price <= @maxPrice
+        AND property.price >= @minPrice
+        AND property.size <= @maxSize
+        AND property.size >= @minSize
+        AND property.parkings >= @parkingCounts
+        AND PropertyType.nom LIKE '%' + @propertyTypes + '%'
+        AND (property.title LIKE '%' + @searchText + '%' OR property.descr LIKE '%' + @searchText + '%')`;
+    } else {
+      query = `
+    SELECT
+        property.*,
+        city.*,
+        PropertyType.*,
+        CASE WHEN Likes.propertyId IS NOT NULL THEN 1 ELSE 0 END AS is_liked
+    FROM
+        property
+    JOIN
+        city ON property.cityId = city.cityId
+    JOIN
+        PropertyType ON property.propertyTypeId = PropertyType.propertyTypeId
+    WHERE
+        property.addres LIKE '%' + @address + '%'
+        AND property.baths > @bathroomCount
+        AND property.beds > @bedroomCount
+        AND city.nom LIKE '%' + @cities + '%'
+        AND property.kitchens > @kitchenRoomCount
+        AND property.price <= @maxPrice
+        AND property.price >= @minPrice
+        AND property.size <= @maxSize
+        AND property.size >= @minSize
+        AND property.parkings >= @parkingCounts
+        AND PropertyType.nom LIKE '%' + @propertyTypes + '%'
+        AND (property.title LIKE '%' + @searchText + '%' OR property.descr LIKE '%' + @searchText + '%')`;
+    }
+
+    const result = await pool
+      .request()
+      .input("user", sql.Int, userId)
+      .input("address", sql.VarChar, address)
+      .input("bathroomCount", sql.Int, bathroomCount)
+      .input("bedroomCount", sql.Int, bedroomCount)
+      .input("cities", sql.VarChar, cities)
+      .input("kitchenRoomCount", sql.Int, kitchenRoomCount)
+      .input("maxPrice", sql.Float, maxPricee)
+      .input("maxSize", sql.Float, maxSizee)
+      .input("minPrice", sql.Float, minPricee)
+      .input("minSize", sql.Float, minSizee)
+      .input("parkingCounts", sql.Int, parkingCountse)
+      .input("propertyTypes", sql.VarChar, propertyTypese)
+      .input("searchText", sql.VarChar, searchTexte)
+      .query(query);
+    const searchProperties = result.recordset; // Check if the user exists
+
+    const response = searchProperties.map((property) => mapProperty(property));
+
+    return res.json(response);
+  }
+});
+
+function mapProperty(property) {
+  // Map the property fields to your Property model
+  return {
+    // Map the fields accordingly
+    // Example:
+    id: property.propertyId,
+    title: property.title,
+    description: property.descr,
+    city: "Kinshasa",
+    imageNames: property.images,
+    bedRoomCount: property.beds,
+    bathRoomCount: property.baths,
+    kitchenRoomCount: property.kitchens,
+    parkingCount: property.parkings,
+    price: property.price,
+    address: property.addres,
+    latitude: property.latitude,
+    longitude: property.longitude,
+    datePosted: property.datePosted,
+    size: property.size,
+    propertyType: property.nom,
+    propertyCategory: property.saleType,
+    user: property.name,
+    agencyAddress: property.adress,
+    agencyEmail: property.email,
+    isLiked: property.is_liked,
+  };
+}
 
 function decryptToken(token) {
   try {
